Use `map` instead of `andThen`
The `sexDecoder` is more complex than it needs to be. Because it's using
`andThen`, it needs to manually re-wrap the values using
`Json.Decode.succeed`.

At a conceptual level, the `sexDecoder` is transforming a string to
`Sex` via a mapping. This is a perfect use-case for the simpler
`Json.Decode.map` function.